Reviewed-on: https://git.vepay.cf/backend/processing/pulls/1020

+++ services/payment/payment_strategies/CreatePayStrategy.php
@@ -252,11 +252,12 @@ class CreatePayStrategy
     public function releaseLock()
     {
         $paySchet = $this->createPayForm->getPaySchet();
+        if ($paySchet) {
+            Yii::info("CreatePayStrategy releaseLock PaySchet.ID={$paySchet->ID}");
 
-        Yii::info("CreatePayStrategy releaseLock PaySchet.ID={$paySchet->ID}");
-
-        $cacheKey = $this->getCacheKey($paySchet);
-        Yii::$app->cache->delete($cacheKey);
+            $cacheKey = $this->getCacheKey($paySchet);
+            Yii::$app->cache->delete($cacheKey);
+        }
     }
 
     /**
